requirements.txt
# Core dependencies
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
click==8.1.7

# Scheduling
python-crontab==3.0.0
croniter==2.0.1

# Notifications
discord-webhook==1.1.0
pushover-complete==1.1.1